Advent of Code 2020, Day 20 Part 2
Adds the code to line up tiles together and then build the image
around it. There's a bit of randomness involved in which way it
will orient the first corner and one of the two orientations
ends up being in the 'right' orientation to solve the puzzle
for our input, so it is not fully 'functional' for other inputs
likely.

https://adventofcode.com/2020/day/20